feat(xo6/PoolMenu): update architecture
