--- @xen-orchestra/lite/src/components/host/HostMenu.vue
@@ -1,36 +1,0 @@
-<template>
-  <MenuList placement="bottom-start">
-    <template #trigger="{ open, isOpen }">
-      <UiButtonIcon
-        accent="brand"
-        icon="action:more-actions"
-        variant="tertiary"
-        size="small"
-        :selected="isOpen"
-        @click="open($event)"
-      />
-    </template>
-    <MenuItem icon="action:download" @click="download()">{{ t('action:download-bugtools-archive') }}</MenuItem>
-  </MenuList>
-</template>
-
-<script lang="ts" setup>
-import MenuItem from '@core/components/menu/MenuItem.vue'
-import MenuList from '@core/components/menu/MenuList.vue'
-import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
-import type { XenApiHost } from '@vates/types'
-import { useI18n } from 'vue-i18n'
-
-const { hostIp } = defineProps<{ hostIp: XenApiHost['address'] }>()
-
-const { t } = useI18n()
-
-const download = () => {
-  const link = document.createElement('a')
-  link.href = `http://${hostIp}/system-status?output=tar.bz2`
-  link.download = 'system-status.tar.bz2'
-  document.body.appendChild(link)
-  link.click()
-  document.body.removeChild(link)
-}
-</script>
