+++ @xen-orchestra/lite/src/modules/host/components/actions/HostTreeActions.vue
@@ -1,0 +1,12 @@
+<template>
+  <HostMoreActions :host />
+</template>
+
+<script setup lang="ts">
+import type { XenApiHost } from '@/libs/xen-api/xen-api.types'
+import HostMoreActions from '@/modules/host/components/actions/HostMoreActions.vue'
+
+defineProps<{
+  host: XenApiHost
+}>()
+</script>
