+++ @xen-orchestra/lite/src/modules/pool/components/actions/PoolMoreActions.vue
@@ -1,0 +1,12 @@
+<template>
+  <PoolDownloadButton :master-host="pool.master" />
+</template>
+
+<script setup lang="ts">
+import type { XenApiPool } from '@/libs/xen-api/xen-api.types.ts'
+import PoolDownloadButton from '@/modules/pool/components/actions/download/PoolDownloadButton.vue'
+
+defineProps<{
+  pool: XenApiPool
+}>()
+</script>
