+++ @xen-orchestra/lite/src/modules/pool/components/actions/poolTreeActions.vue
@@ -1,0 +1,12 @@
+<template>
+  <PoolMoreActions :pool />
+</template>
+
+<script setup lang="ts">
+import type { XenApiPool } from '@/libs/xen-api/xen-api.types.ts'
+import PoolMoreActions from '@/modules/pool/components/actions/PoolMoreActions.vue'
+
+defineProps<{
+  pool: XenApiPool
+}>()
+</script>
