+++ @xen-orchestra/lite/typed-router.d.ts
@@ -62,6 +62,7 @@ declare module 'vue-router/auto-routes' {
     '/story/web-core/input-wrapper/vts-input-wrapper': RouteRecordInfo<'/story/web-core/input-wrapper/vts-input-wrapper', '/story/web-core/input-wrapper/vts-input-wrapper', Record<never, never>, Record<never, never>>,
     '/story/web-core/legend-group/vts-legend-group': RouteRecordInfo<'/story/web-core/legend-group/vts-legend-group', '/story/web-core/legend-group/vts-legend-group', Record<never, never>, Record<never, never>>,
     '/story/web-core/object-icon/vts-object-icon': RouteRecordInfo<'/story/web-core/object-icon/vts-object-icon', '/story/web-core/object-icon/vts-object-icon', Record<never, never>, Record<never, never>>,
+    '/story/web-core/resources/vts-resource': RouteRecordInfo<'/story/web-core/resources/vts-resource', '/story/web-core/resources/vts-resource', Record<never, never>, Record<never, never>>,
     '/story/web-core/resources/vts-resources': RouteRecordInfo<'/story/web-core/resources/vts-resources', '/story/web-core/resources/vts-resources', Record<never, never>, Record<never, never>>,
     '/story/web-core/select/vts-select': RouteRecordInfo<'/story/web-core/select/vts-select', '/story/web-core/select/vts-select', Record<never, never>, Record<never, never>>,
     '/story/web-core/stacked-bar-with-legend/vts-stacked-bar-with-legend': RouteRecordInfo<'/story/web-core/stacked-bar-with-legend/vts-stacked-bar-with-legend', '/story/web-core/stacked-bar-with-legend/vts-stacked-bar-with-legend', Record<never, never>, Record<never, never>>,
@@ -70,6 +71,7 @@ declare module 'vue-router/auto-routes' {
     '/story/web-core/tab/router-tab': RouteRecordInfo<'/story/web-core/tab/router-tab', '/story/web-core/tab/router-tab', Record<never, never>, Record<never, never>>,
     '/story/web-core/tab/tab-item': RouteRecordInfo<'/story/web-core/tab/tab-item', '/story/web-core/tab/tab-item', Record<never, never>, Record<never, never>>,
     '/story/web-core/tab/tab-list': RouteRecordInfo<'/story/web-core/tab/tab-list', '/story/web-core/tab/tab-list', Record<never, never>, Record<never, never>>,
+    '/story/web-core/tree/vts-tree-item': RouteRecordInfo<'/story/web-core/tree/vts-tree-item', '/story/web-core/tree/vts-tree-item', Record<never, never>, Record<never, never>>,
     '/story/web-core/tree/vts-tree-item-error': RouteRecordInfo<'/story/web-core/tree/vts-tree-item-error', '/story/web-core/tree/vts-tree-item-error', Record<never, never>, Record<never, never>>,
     '/story/web-core/tree/vts-tree-list': RouteRecordInfo<'/story/web-core/tree/vts-tree-list', '/story/web-core/tree/vts-tree-list', Record<never, never>, Record<never, never>>,
     '/story/web-core/tree/vts-tree-loading-item': RouteRecordInfo<'/story/web-core/tree/vts-tree-loading-item', '/story/web-core/tree/vts-tree-loading-item', Record<never, never>, Record<never, never>>,
