+++ @xen-orchestra/web-core/lib/locales/en.json
@@ -32,6 +32,7 @@
   "action:delete-vms": "Delete 1 VM | Delete {n} VMs",
   "action:deploy": "Deploy",
   "action:deploy-xoa": "Deploy XOA",
+  "action:download-bugtools-archive": "Download bugtools archive",
   "action:duplicate": "Duplicate",
   "action:edit": "Edit",
   "action:edit-config": "Edit config",
@@ -277,7 +278,6 @@
   "documentation": "Documentation",
   "documentation-name": "{name} documentation",
   "done": "Done",
-  "action:download-bugtools-archive": "Download bugtools archive",
   "duration": "Duration",
   "enabled": "Enabled",
   "end-date": "End date",
