+++ @xen-orchestra/web-core/lib/locales/fr.json
@@ -32,6 +32,7 @@
   "action:delete-vms": "Supprimer 1 VM | Supprimer {n} VMs",
   "action:deploy": "Déployer",
   "action:deploy-xoa": "Déployer XOA",
+  "action:download-bugtools-archive": "Télécharger l’archive bugtools",
   "action:duplicate": "Dupliquer",
   "action:edit": "Modifier",
   "action:edit-config": "Modifier config",
@@ -277,7 +278,6 @@
   "documentation": "Documentation",
   "documentation-name": "Documentation {name}",
   "done": "Terminé",
-  "action:download-bugtools-archive": "Télécharger l’archive bugtools",
   "duration": "Durée",
   "enabled": "Activé",
   "end-date": "Date de fin",
