+++ @xen-orchestra/web/src/modules/pool/components/actions/PoolTreeActions.vue
@@ -1,0 +1,12 @@
+<template>
+  <PoolMoreActions :pool />
+</template>
+
+<script setup lang="ts">
+import PoolMoreActions from '@/modules/pool/components/actions/PoolMoreActions.vue'
+import type { XoPool } from '@vates/types'
+
+defineProps<{
+  pool: XoPool
+}>()
+</script>
