+++ @xen-orchestra/web/src/modules/treeview/components/HostTreeItem.vue
@@ -19,7 +19,19 @@
           variant="secondary"
           size="small"
         />
-        <HostMenu :host-ip="branch.data.address" />
+        <MenuList placement="bottom-start">
+          <template #trigger="{ open, isOpen }">
+            <UiButtonIcon
+              accent="brand"
+              icon="action:more-actions"
+              variant="tertiary"
+              size="small"
+              :selected="isOpen"
+              @click="open($event)"
+            />
+          </template>
+          <HostTreeActions :host="branch.data" />
+        </MenuList>
       </template>
     </UiTreeItemLabel>
     <template v-if="branch.hasChildren" #sublist>
@@ -31,16 +43,18 @@
 </template>
 
 <script lang="ts" setup>
-import HostMenu from '@/modules/host/components/HostMenu.vue'
+import HostTreeActions from '@/modules/host/components/actions/HostTreeActions.vue'
 import { useXoHostCollection } from '@/modules/host/remote-resources/use-xo-host-collection.ts'
 import VmTreeList from '@/modules/treeview/components/VmTreeList.vue'
 import type { HostBranch } from '@/modules/treeview/types/tree.type.ts'
 import { useXoVmCollection } from '@/modules/vm/remote-resources/use-xo-vm-collection.ts'
 import type { HostState } from '@core/types/object-icon.type.ts'
 import VtsIcon from '@core/components/icon/VtsIcon.vue'
+import MenuList from '@core/components/menu/MenuList.vue'
 import VtsObjectIcon from '@core/components/object-icon/VtsObjectIcon.vue'
 import VtsTreeItem from '@core/components/tree/VtsTreeItem.vue'
 import VtsTreeList from '@core/components/tree/VtsTreeList.vue'
+import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
 import UiCounter from '@core/components/ui/counter/UiCounter.vue'
 import UiTreeItemLabel from '@core/components/ui/tree-item-label/UiTreeItemLabel.vue'
 import { vTooltip } from '@core/directives/tooltip.directive.ts'
